backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
pydantic-settings==2.13.1
sqlalchemy==2.0.48
asyncpg==0.31.0
python-jose[cryptography]==3.5.0
httpx==0.28.1
python-multipart==0.0.22
google-genai>=1.0.0
Pillow>=10.0.0
